Change the light prob distribution
